Add support for dind
Added handling to the server_running method to support cases where
the docker container is running testcontainers inside a container.
When this flow is used users will need to set an WIREMOCK_DIND
env var so that the host used can revert to the host.docker.internal
url.

There are also a number of fixes to the code to get the test suites
passing.  Test run time was all reduced significantly by mocking
more method to avoid starting containers on a test by test bases.